+++ CHANGES.rst
@@ -19,8 +19,8 @@ Release History
    - Deprecated
    - Removed
 
-2.2.1 (unreleased)
-==================
+2.2.1 (October 2, 2019)
+=======================
 
 **Changed**